feat: moving logic to service

+++ Tennis-scheduler/src/main/java/com/tennis/tennisscheduler/service/TimeslotService.java
@@ -1,16 +1,21 @@
 package com.tennis.tennisscheduler.service;
 
+import com.tennis.tennisscheduler.exception.ApiRequestException;
+import com.tennis.tennisscheduler.model.Person;
 import com.tennis.tennisscheduler.model.Timeslot;
+import com.tennis.tennisscheduler.model.enumes.UserType;
 import com.tennis.tennisscheduler.repository.TimeslotRepository;
 import com.tennis.tennisscheduler.response.TimeslotResponse;
+import com.tennis.tennisscheduler.security.SecurityContextUtil;
 import lombok.AllArgsConstructor;
+import org.springframework.http.HttpStatus;
 import org.springframework.scheduling.annotation.Scheduled;
 import org.springframework.stereotype.Service;
+import org.springframework.validation.BindingResult;
 
 import javax.persistence.EntityNotFoundException;
 import java.util.Date;
 import java.util.List;
-import java.util.Optional;
 
 @Service
 @AllArgsConstructor
@@ -19,21 +24,13 @@ public class TimeslotService {
     private final PersonService personService;
     private final TennisCourtService tennisCourtService;
 
-    public TimeslotResponse update(long id, Timeslot timeslot){
-        Timeslot existingTimeslot = timeslotRepository.findById(id).get();
+    public TimeslotResponse update(long id, Timeslot timeslot, BindingResult result){
+        Timeslot existingTimeslot = timeslotRepository.findById(id).orElseThrow(EntityNotFoundException::new);
         existingTimeslot.setStartDate(timeslot.getStartDate());
         existingTimeslot.setEndDate(timeslot.getEndDate());
         existingTimeslot.setPerson(personService.findById(timeslot.getPerson().getId()).get());
         existingTimeslot.setTennisCourt(tennisCourtService.getTennisCourtById(timeslot.getTennisCourt().getId()).get());
-        return reserveTimeslot(existingTimeslot);
-    }
-
-    public List<Timeslot> getAll(){
-        return timeslotRepository.getALlTimeslotsAdmin();
-    }
-
-    public List<Timeslot> getAllTimeslotsForUser(long personId){
-        return timeslotRepository.getAllTimeslotsForUser(personId);
+        return reserveTimeslot(existingTimeslot, null);
     }
 
     public Timeslot getById(Long id){
@@ -41,6 +38,10 @@ public class TimeslotService {
     }
 
     public void deleteById(long id){
+        Timeslot existingTimeslot = timeslotRepository.findById(id).orElseThrow(EntityNotFoundException::new);
+        if (SecurityContextUtil.GetLoggedUser().getRole().getRoleName().equals(UserType.ROLE_TENNIS_PLAYER) && existingTimeslot.getPerson().getId() != SecurityContextUtil.GetLoggedUser().getId())
+            throw new ApiRequestException(HttpStatus.UNAUTHORIZED,"You don't have permission");
+
         timeslotRepository.deleteById(id);
     }
 
@@ -51,15 +52,30 @@ public class TimeslotService {
         }
     }
 
-    public Timeslot save(Timeslot timeslot) {
+    private Timeslot save(Timeslot timeslot) {
         timeslot.setTennisCourt(tennisCourtService.getTennisCourtById(timeslot.getTennisCourt().getId()).get());
         timeslot.setPerson(personService.findById(timeslot.getPerson().getId()).get());
         return timeslotRepository.save(timeslot);
     }
 
-    public TimeslotResponse reserveTimeslot(Timeslot timeslot) {
+    public TimeslotResponse reserveTimeslot(Timeslot timeslot, BindingResult result) {
+        Person user = SecurityContextUtil.GetLoggedUser();
         TimeslotResponse timeslotResponse = new TimeslotResponse();
+
+        if (!result.hasErrors() && user.getRole().getRoleName().equals(UserType.ROLE_TENNIS_PLAYER))
+            timeslot.setPerson(user);
+
+        if (result.hasErrors())
+            throw new ApiRequestException(HttpStatus.BAD_REQUEST,"BAD REQUEST");
+
         timeslotResponse.setTimeslot(save(timeslot));
         return timeslotResponse;
+    }
+
+    public List<Timeslot> getAllTimeslots() {
+        if(SecurityContextUtil.GetLoggedUser().getRole().getRoleName().equals(UserType.ROLE_TENNIS_PLAYER))
+            return timeslotRepository.getAllTimeslotsForUser(SecurityContextUtil.GetLoggedUser().getId());
+        else
+            return timeslotRepository.findAll();
     }
 }
